[refactor] Use camelCase for transactionSubmission
