+++ src/api/aptos.ts
@@ -12,7 +12,7 @@ import { General } from "./general";
 import { Staking } from "./staking";
 import { Token } from "./token";
 import { Transaction } from "./transaction";
-import { TransactionSubmission } from "./transaction_submission";
+import { TransactionSubmission } from "./transactionSubmission";
 
 /**
  * This class is the main entry point into Aptos's
